Changed to a simple cmd app.
This fork was created as an accident in the first place, so I didn't really need it and it wasn't mainteined properly. The original repo has a lot more work on UI side, that this repo lacked, so you might want to check out the original instead. In this commit all the unfinished pieces are removed and now this fork contains only the fuzzy-logic inference logic, a simple console app to check it and some docs.